Improves display of show pages
Refactors show pages for admins, categories, products and subcategories to use a more structured and visually appealing layout.

Replaces the previous grid-based layout with a stack-based layout utilizing Paper and Box components for better organization.

Introduces an InfoField component for displaying record details with labels, and a SectionTitle component to improve readability.

Sets default status to "DRAFT" when creating new categories and products.

Sets lastEditedBy field to the current user when creating categories and products.